refactor(tests): move the Flash-shaped kernel tests into their own module
test_deepseek_v4.py held two suites. Above the boundary a toy five-layer _MODEL covers checkpoint
conversion, vLLM name mapping, RoPE and the packing invariants; below it a second config at the
real V4-Flash shapes existed only because the sparse path is invisible at toy sizes. More kernels
are coming, so the lower half moves to the file that already owns kernel coverage.

Nothing in the block changed: this is a cut and paste, with the imports and shared helpers landed
separately beforehand so the move touches only line positions. `copy`, `requires_tilelang` and
_set_attn_impl go with it, having no reader left behind.
