+++ tests/unit/train/models/test_deepseek_v4.py
@@ -7,7 +7,6 @@ archetype the other models in this directory use, where a tiny `HF<X>ForCausalLM
 expected logits and gradients.
 """
 
-import copy
 import math
 import re
 from unittest.mock import MagicMock
@@ -27,12 +26,6 @@ from prime_rl.trainer.models.layers.lm_head import inject_prime_lm_head
 from prime_rl.utils.utils import default_dtype
 
 pytestmark = [pytest.mark.gpu]
-
-# The fused kernel needs tilelang, which only the `gpu` extra provides and only on linux. Most of
-# this module runs without it, so this is a per-test skip rather than part of `pytestmark`.
-requires_tilelang = pytest.mark.skipif(
-    dsv4_attention.dsv4_sparse_attn is None, reason="the sparse attention kernel needs tilelang"
-)
 
 # Deliberately heterogeneous: one layer of every attention type, hash-routed bootstrap
 # layers ahead of standard MoE ones, and a sliding window narrow enough that the compressed
@@ -164,17 +157,6 @@ def prime_attention(layer_idx: int, dtype: torch.dtype = torch.bfloat16, dsv4_at
         module = DeepseekV4Attention(_prime_config(dsv4_attn), layer_idx=layer_idx)
     _randomize(module)
     return module
-
-
-def _set_attn_impl(module: nn.Module, impl: str) -> None:
-    """Pin the CSA attention implementation on every attention layer `module` owns.
-
-    `modules()` yields `module` itself, so this covers a lone attention layer as well as a model
-    holding several of them.
-    """
-    for submodule in module.modules():
-        if isinstance(submodule, DeepseekV4Attention):
-            submodule.attn_impl = impl
 
 
 def _single_doc(input_ids: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
@@ -951,491 +933,3 @@ def test_attention_packed_matches_unpacked(layer_idx, doc_lens, selection, _torc
 
     _compare_accumulated_grads(module, packed_grads)
     torch.testing.assert_close(alone_input.grad, packed_input.grad, rtol=_PACKED_RTOL, atol=_PACKED_ATOL)
-
-
-# Everything below runs the real DeepSeek V4 Flash attention shapes instead of the toy `_MODEL`
-# above, written out as a literal so nothing here depends on a local HF cache. The sparse gather
-# path only exists at those shapes: a CSA query reads `sliding_window + index_topk = 640` keys
-# over 512 channels with 64 heads, and the slot padding, the top-k saturation and the index
-# arithmetic that the sparse representation has to get right are all invisible at toy sizes. The
-# MoE fields are shrunk to nothing, since `DeepseekV4Attention` reads none of them.
-_FLASH_MODEL = dict(
-    vocab_size=64,
-    hidden_size=4096,
-    num_attention_heads=64,
-    num_key_value_heads=1,
-    head_dim=512,
-    q_lora_rank=1024,
-    o_groups=8,
-    o_lora_rank=1024,
-    qk_rope_head_dim=64,
-    rope_theta=10000.0,
-    compress_rope_theta=160000.0,
-    sliding_window=128,
-    index_n_heads=64,
-    index_head_dim=128,
-    index_topk=512,
-    compress_rates={"compressed_sparse_attention": 4, "heavily_compressed_attention": 128},
-    layer_types=["compressed_sparse_attention", "heavily_compressed_attention"],
-    num_hidden_layers=2,
-    rms_norm_eps=1e-6,
-    attention_dropout=0.0,
-    max_position_embeddings=65536,
-    moe_intermediate_size=64,
-    n_routed_experts=8,
-    num_experts_per_tok=3,
-    n_shared_experts=1,
-    scoring_func="sqrtsoftplus",
-    routed_scaling_factor=1.5,
-    swiglu_limit=10.0,
-    num_hash_layers=1,
-    hc_mult=4,
-    hc_sinkhorn_iters=20,
-    hc_eps=1e-6,
-    rope_scaling={
-        "beta_fast": 32,
-        "beta_slow": 1,
-        "factor": 16,
-        "original_max_position_embeddings": 65536,
-        "type": "yarn",
-    },
-)
-
-_FLASH_CSA_LAYER = 0
-_FLASH_COMPRESS_RATE = _FLASH_MODEL["compress_rates"]["compressed_sparse_attention"]
-
-# Document layouts for the sparse path, at `compress_rate = 4`. The first four leave every query
-# short of `index_topk = 512` readable entries, so the `-1` padding of the pick slots carries
-# the difference; `(2600,)` saturates the picks instead, which the toy shapes cannot express at
-# all. `(3,)` compresses to no entries whatsoever, leaving the local window alone to answer.
-_FLASH_DOC_LENS = [(517, 1019), (3,), (300,), (3, 129, 1021), (2600,)]
-_FLASH_DOC_IDS = ["two-docs", "no-entries", "one-short-doc", "three-docs", "saturated-topk"]
-
-# The two paths sum their weighted values in a different order (640 gathered slots against
-# `seq_len + n_entries` masked columns) over 512 channels, so they agree only to the float32
-# accumulation floor rather than bit for bit. Measured over the five layouts below, against each
-# tensor's own scale, the worst deviation is 2.5e-7 on the output and 1.8e-6 on a gradient, so
-# `_PACKED_RTOL` still holds here with room to spare and neither number is tuned to fit.
-_GATHER_RTOL, _GATHER_GRAD_RTOL = _PACKED_RTOL, _PACKED_GRAD_RTOL
-
-
-def _flash_config(dsv4_attn: str = "auto") -> DeepseekV4Config:
-    return DeepseekV4Config(**_FLASH_MODEL, dsv4_attn=dsv4_attn)
-
-
-def flash_attention(layer_idx: int, dtype: torch.dtype = torch.float32, dsv4_attn: str = "auto") -> nn.Module:
-    """One attention layer at the real DeepSeek V4 Flash shapes, 126M parameters of it."""
-    with torch.device("cuda"), default_dtype(dtype):
-        module = DeepseekV4Attention(_flash_config(dsv4_attn), layer_idx=layer_idx)
-    _randomize(module)
-    return module
-
-
-def _flash_hidden_states(seq_len: int) -> tuple[torch.Tensor, torch.Tensor]:
-    """Two leaves carrying identical values, one per attention path.
-
-    Batch 1: at these shapes the score tensors are the bulk of the memory and a second batch
-    entry repeats the first without covering anything new.
-    """
-    with torch.device("cuda"):
-        hidden = torch.randn(1, seq_len, _FLASH_MODEL["hidden_size"])
-    return hidden.clone().requires_grad_(True), hidden.clone().requires_grad_(True)
-
-
-def _record_attention(monkeypatch) -> dict[str, torch.Tensor]:
-    """Capture what each attention path is handed and what it returns, from real forwards.
-
-    Both implementations are module-level functions the layer looks up by name, so patching them
-    records the call the layer actually made rather than a reconstruction of it.
-    """
-    recorded: dict[str, torch.Tensor] = {}
-    real_eager = dsv4_attention.eager_attention_with_sinks
-    real_gather = dsv4_attention.sparse_attention_gather
-
-    def eager(query, key, value, sinks, attention_mask, **kwargs):
-        recorded["mask"] = attention_mask
-        recorded["eager"] = real_eager(query, key, value, sinks, attention_mask, **kwargs)
-        return recorded["eager"]
-
-    def gather(q, kv_buf, indices, sinks, scale):
-        recorded["kv_buf"], recorded["indices"] = kv_buf, indices
-        recorded["gather"] = real_gather(q, kv_buf, indices, sinks, scale)
-        return recorded["gather"]
-
-    monkeypatch.setattr(dsv4_attention, "eager_attention_with_sinks", eager)
-    monkeypatch.setattr(dsv4_attention, "sparse_attention_gather", gather)
-    return recorded
-
-
-def _selected_positions(indices: torch.Tensor, n_positions: int) -> torch.Tensor:
-    """`(seq_len, n_positions)` bool: which KV positions each query's gather slots address."""
-    slots = indices[0, :, 0, :].long()
-    # `-1` marks an absent key; it goes into one throwaway column that is sliced back off.
-    safe = torch.where(slots >= 0, slots, n_positions)
-    selected = torch.zeros((indices.shape[1], n_positions + 1), dtype=torch.bool, device=indices.device)
-    return selected.scatter_(1, safe, True)[:, :n_positions]
-
-
-@pytest.mark.parametrize("doc_lens", _FLASH_DOC_LENS, ids=_FLASH_DOC_IDS)
-def test_sparse_indices_address_exactly_the_keys_the_dense_mask_admits(doc_lens, monkeypatch):
-    """The two CSA paths must reach the same keys, key for key, for every query.
-
-    They render one selection two independent ways: the dense path concatenates the indexer's
-    picks onto the sliding mask, the sparse path writes the window and the picks into one index
-    tensor over a gathered KV buffer. Nothing else in the layer compares them, and every way of
-    getting the sparse side wrong (a window base off by one, an entry index not offset by the
-    token count, a stale index left over from a previous layout) still produces a finite output.
-
-    Pure set equality on integers, so no tolerance enters. Both representations are captured from
-    real forwards of the same module on the same input.
-    """
-    module = flash_attention(_FLASH_CSA_LAYER)
-    packed = _packed_context(doc_lens, torch.float32, _flash_config())
-    hidden_states = _flash_hidden_states(sum(doc_lens))[0].detach()
-    recorded = _record_attention(monkeypatch)
-
-    with torch.no_grad():
-        _set_attn_impl(module, "eager")
-        module(hidden_states, packed=packed)
-        _set_attn_impl(module, "gather")
-        module(hidden_states, packed=packed)
-
-    admitted = recorded["mask"][0, 0] == 0  # (seq_len, seq_len + n_entries)
-    n_positions = recorded["kv_buf"].shape[1]
-    seq_len, n_entries = sum(doc_lens), n_positions - sum(doc_lens)
-    assert n_entries == sum(length // _FLASH_COMPRESS_RATE for length in doc_lens)
-    if n_entries:
-        assert admitted[:, seq_len:].any(), "vacuous probe: no query reads a compressed entry"
-
-    selected = _selected_positions(recorded["indices"], n_positions)
-    assert torch.equal(selected, admitted), "the sparse and dense paths select different keys"
-
-
-@pytest.mark.parametrize("doc_lens", _FLASH_DOC_LENS, ids=_FLASH_DOC_IDS)
-def test_sparse_indices_are_in_range_and_never_repeat_a_key(doc_lens, monkeypatch):
-    """Every gather slot addresses a real KV position, and no query counts a key twice.
-
-    A negative or out-of-range index reads whatever lies next to the buffer instead of raising,
-    which the kernel that will consume these has no way to detect. A repeat is worse than
-    wasteful: the duplicated key takes twice its share of the softmax, silently reweighting the
-    output. The `-1` padding is exempt from uniqueness, since padding every query out to a fixed
-    slot count is exactly what it is for.
-    """
-    module = flash_attention(_FLASH_CSA_LAYER, dsv4_attn="gather")
-    packed = _packed_context(doc_lens, torch.float32, _flash_config())
-    hidden_states = _flash_hidden_states(sum(doc_lens))[0].detach()
-    recorded = _record_attention(monkeypatch)
-
-    with torch.no_grad():
-        module(hidden_states, packed=packed)
-
-    indices, n_positions = recorded["indices"], recorded["kv_buf"].shape[1]
-    # The slot count is `sliding_window + index_topk` rounded up to the kernel's slot tile.
-    # The Flash shapes need no rounding, 640 being a multiple of 64 already, so the padding
-    # itself is covered by the toy `_MODEL` under `dsv4_attn='gather'` and not here.
-    n_slots = indices.shape[-1]
-    assert n_slots >= _FLASH_MODEL["sliding_window"] + _FLASH_MODEL["index_topk"]
-    assert n_slots % dsv4_attention._SLOT_TILE == 0
-    assert (indices >= -1).all(), "a gather slot addresses a KV position below the `-1` marker"
-    assert (indices <= n_positions - 1).all(), "a gather slot addresses past the end of the KV buffer"
-
-    slot_idx = indices[0, :, 0, :].long()
-    # `-1` is counted in a throwaway column, since padding repeats it by design.
-    safe = torch.where(slot_idx >= 0, slot_idx, n_positions)
-    counts = torch.zeros((slot_idx.shape[0], n_positions + 1), dtype=torch.int32, device="cuda")
-    counts.scatter_add_(1, safe, torch.ones_like(safe, dtype=torch.int32))
-    assert (counts[:, :n_positions] <= 1).all(), "a query gathers the same key twice"
-
-
-@pytest.mark.parametrize("doc_lens", _FLASH_DOC_LENS, ids=_FLASH_DOC_IDS)
-def test_sparse_attention_gather_matches_eager(doc_lens, monkeypatch):
-    """The gather reference must reproduce the dense path's attention output and its gradients.
-
-    This is the contract the TileLang kernel will be validated against, so it is asserted on the
-    raw `(batch, seq_len, heads, head_dim)` attention output rather than through `o_a_proj` and
-    `o_b_proj`, which would add projection error without covering anything the kernel does. Both
-    halves run the same weights on the same values, one leaf each, with the implementation pinned
-    explicitly on the built module so neither `dsv4_attn` nor `auto` can collapse them onto one
-    path.
-
-    The backward matters as much as the forward: the sink logit is the one term the gather
-    reference handles differently (unscaled, concatenated, then stripped back off), and a sink
-    that never receives a gradient would leave the softmax denominator untrained.
-    """
-    module = flash_attention(_FLASH_CSA_LAYER)
-    packed = _packed_context(doc_lens, torch.float32, _flash_config())
-    eager_input, gather_input = _flash_hidden_states(sum(doc_lens))
-    recorded = _record_attention(monkeypatch)
-
-    _set_attn_impl(module, "eager")
-    module(eager_input, packed=packed)
-    eager_output = recorded["eager"]
-
-    _set_attn_impl(module, "gather")
-    module(gather_input, packed=packed)
-    gather_output = recorded["gather"]
-
-    assert gather_output.shape == eager_output.shape
-    _assert_relative(gather_output, eager_output, _GATHER_RTOL, "attention output")
-
-    # One weight tensor for both losses, so the two backwards are the same function of the same
-    # numbers and any difference is the attention path's alone.
-    with torch.device("cuda"):
-        weight = torch.randn_like(eager_output)
-    (eager_output * weight).sum().backward()
-    eager_grads = _take_grads(module)
-    assert eager_grads["sinks"] is not None and eager_grads["sinks"].norm() > 0, (
-        "vacuous probe: the sinks received no gradient, so the comparison below cannot fail on them"
-    )
-
-    (gather_output * weight).sum().backward()
-    _compare_accumulated_grads(module, eager_grads, rtol=_GATHER_GRAD_RTOL)
-    _assert_relative(gather_input.grad, eager_input.grad, _GATHER_GRAD_RTOL, "hidden states gradient")
-
-
-# One CSA layer in bfloat16, so `_PACKED_RTOL` (float32, and three orders of magnitude tighter
-# than a kernel accumulating bfloat16 inputs) does not apply, but neither does `_MODEL_GRAD_RTOL`,
-# which is sized for four hyper-connected layers amplifying a bf16 expert floor. Each bound below
-# is the tightest round number holding over 30 seeds; the worst is 1.4e-3 on the output and
-# 7.6e-3 on a gradient, against 6.9e-4 and 6.2e-3 on the fixed seed the test actually runs. The
-# gradient bound is the tighter fit of the two, at 1.3x: every seed lands between 5.8e-3 and
-# 7.6e-3, so the bound sits just above a well-sampled ceiling rather than above a long tail.
-_KERNEL_RTOL, _KERNEL_GRAD_RTOL = 5e-3, 1e-2
-
-# `compress_rate = 4` yields 129 + 254 = 383 compressed entries, under `index_topk = 512`, so
-# every readable entry is picked and the indexer's ordering cannot differ packed from alone. A
-# saturated layout would let a bfloat16 tie flip a pick and move the output for a reason that has
-# nothing to do with document independence.
-_KERNEL_DOC_LENS = (517, 1019)
-
-
-# Document layouts for the kernel-against-eager comparison: two single-document rows and two
-# packed ones. `(2600,)` is left out on purpose for the reason `_KERNEL_DOC_LENS` gives below, and
-# `(3,)` is kept because it compresses to no entries at all, so almost every gather slot is the
-# `-1` marker and the local window alone has to answer.
-_EAGER_KERNEL_DOC_LENS = [(300,), (3,), (517, 1019), (3, 129, 1021)]
-_EAGER_KERNEL_DOC_IDS = ["one-doc", "no-entries", "two-docs", "three-docs"]
-
-# A bfloat16 kernel against a float32 dense softmax, so these are three orders of magnitude looser
-# than the float32 `_GATHER_RTOL` next door and looser again than `_KERNEL_RTOL`, which compares
-# two bfloat16 runs of the same path. Each is the tightest round number holding over 30 seeds on
-# all four layouts: the worst observed is 6.1e-3 on the output and 1.6e-2 on a gradient, against
-# 4.9e-3 and 9.8e-3 on the fixed seed the test runs.
-_EAGER_KERNEL_RTOL, _EAGER_KERNEL_GRAD_RTOL = 8e-3, 2e-2
-
-
-@pytest.mark.parametrize("doc_lens", _EAGER_KERNEL_DOC_LENS, ids=_EAGER_KERNEL_DOC_IDS)
-@requires_tilelang
-def test_sparse_attention_kernel_matches_eager(doc_lens, monkeypatch):
-    """The fused kernel against the naive dense softmax, single-document and packed.
-
-    Every other kernel test reaches eager only transitively: the kernel is compared to the gather
-    reference on hand-built tensors, and the gather reference is compared to eager on a real
-    module. Nothing joined the two ends on the same input, so a disagreement that the gather
-    reference happened to share with the kernel would go unseen. This closes that loop, and it is
-    the only place the modeling code's own index construction meets the kernel in a numeric
-    comparison rather than a set-equality one.
-
-    Both halves hold the same weights, the bfloat16 module's, one widened to float32 rather than
-    drawn again, and both start from the same bfloat16-representable hidden states, so input
-    rounding is not one of the differences being measured. What is left is the attention path:
-    a dense mask and a full softmax on one side, a 640-slot gather and an online softmax on the
-    other.
-
-    The call count is load-bearing rather than decoration: `dsv4_sparse_attn` raises today instead
-    of demoting a dtype it cannot run, but a regression that reintroduced a fallback would leave
-    this comparing eager against eager and passing for the wrong reason.
-    """
-    seq_len = sum(doc_lens)
-    kernel_module = flash_attention(_FLASH_CSA_LAYER, dtype=torch.bfloat16, dsv4_attn="kernel")
-    eager_module = copy.deepcopy(kernel_module).float()
-    _set_attn_impl(eager_module, "eager")
-
-    with torch.device("cuda"):
-        base = torch.randn(1, seq_len, _FLASH_MODEL["hidden_size"], dtype=torch.bfloat16)
-    kernel_input = base.clone().requires_grad_(True)
-    eager_input = base.float().clone().requires_grad_(True)
-
-    calls = []
-    real_kernel = dsv4_attention.dsv4_sparse_attn
-
-    def counting_kernel(q, kv_buf, indices, sinks, scale):
-        calls.append(q.shape[1])
-        return real_kernel(q, kv_buf, indices, sinks, scale)
-
-    monkeypatch.setattr(dsv4_attention, "dsv4_sparse_attn", counting_kernel)
-
-    kernel_output, _ = kernel_module(kernel_input, packed=_packed_context(doc_lens, torch.bfloat16, _flash_config()))
-    eager_output, _ = eager_module(eager_input, packed=_packed_context(doc_lens, torch.float32, _flash_config()))
-    assert calls == [seq_len], f"the forward never reached the kernel, calls={calls}"
-    _assert_relative(kernel_output, eager_output, _EAGER_KERNEL_RTOL, "attention output")
-
-    # One weight tensor for both losses, so any difference belongs to the attention path alone.
-    with torch.device("cuda"):
-        weight = torch.randn(1, seq_len, _FLASH_MODEL["hidden_size"], dtype=torch.float32)
-    (eager_output * weight).sum().backward()
-    eager_grads = _take_grads(eager_module)
-    assert eager_grads["sinks"] is not None and eager_grads["sinks"].norm() > 0, (
-        "vacuous probe: the sinks received no gradient, so the comparison below cannot fail on them"
-    )
-
-    (kernel_output * weight.bfloat16()).sum().backward()
-    _compare_accumulated_grads(kernel_module, eager_grads, rtol=_EAGER_KERNEL_GRAD_RTOL)
-    _assert_relative(kernel_input.grad, eager_input.grad, _EAGER_KERNEL_GRAD_RTOL, "hidden states gradient")
-
-
-@requires_tilelang
-def test_sparse_attention_kernel_packed_matches_unpacked(monkeypatch):
-    """The fused kernel path, end to end through one CSA layer, must respect documents.
-
-    The same invariant its float32 neighbours assert, run in bfloat16 because that is the only
-    dtype `dsv4_sparse_attn` accepts. Numerics belong to
-    `test_dsv4_sparse_attn.py`, which compares the kernel against the float32 gather oracle on
-    hand-built tensors; what is covered here is that the modeling code feeds the kernel inputs it
-    can act on, and that nothing in `q`, the KV buffer or the indices carries the packed row's
-    layout into a document's own answer.
-
-    The call count is load-bearing, not decoration: `dsv4_sparse_attn` raises today rather than
-    demoting a dtype it cannot run, but without counting the calls a regression that reintroduced
-    a fallback would leave this test asserting a property of the gather reference instead.
-    """
-    module = flash_attention(_FLASH_CSA_LAYER, dtype=torch.bfloat16, dsv4_attn="kernel")
-    packed = _packed_context(_KERNEL_DOC_LENS, torch.bfloat16, _flash_config())
-    with torch.device("cuda"):
-        hidden = torch.randn(1, sum(_KERNEL_DOC_LENS), _FLASH_MODEL["hidden_size"], dtype=torch.bfloat16)
-    packed_input, alone_input = hidden.clone().requires_grad_(True), hidden.clone().requires_grad_(True)
-
-    calls = []
-    real_kernel = dsv4_attention.dsv4_sparse_attn
-
-    def counting_kernel(q, kv_buf, indices, sinks, scale):
-        calls.append(q.shape[1])
-        return real_kernel(q, kv_buf, indices, sinks, scale)
-
-    monkeypatch.setattr(dsv4_attention, "dsv4_sparse_attn", counting_kernel)
-
-    packed_output, _ = module(packed_input, packed=packed)
-    assert calls == [sum(_KERNEL_DOC_LENS)], f"the packed forward never reached the kernel, calls={calls}"
-    with torch.device("cuda"):
-        weight = torch.randn_like(packed_output)
-    (packed_output * weight).sum().backward()
-    packed_grads = _take_grads(module)
-
-    for index, length in enumerate(_KERNEL_DOC_LENS):
-        span = _doc_slice(_KERNEL_DOC_LENS, index)
-        alone_output, _ = module(
-            alone_input[:, span], packed=_packed_context((length,), torch.bfloat16, _flash_config())
-        )
-        _assert_relative(packed_output[:, span], alone_output, _KERNEL_RTOL, f"document {index}")
-        (alone_output * weight[:, span]).sum().backward()
-
-    assert calls == [sum(_KERNEL_DOC_LENS), *_KERNEL_DOC_LENS], f"a forward never reached the kernel, calls={calls}"
-    _compare_accumulated_grads(module, packed_grads, rtol=_KERNEL_GRAD_RTOL)
-    _assert_relative(alone_input.grad, packed_input.grad, _KERNEL_GRAD_RTOL, "hidden states gradient")
-
-
-@requires_tilelang
-def test_sparse_attention_kernel_trains_every_parameter(monkeypatch):
-    """Every parameter of a CSA layer that can train does, with the kernel in the path.
-
-    `test_deepseek_v4_backward` makes this assertion through the assembled model, but only on the
-    gather path: the kernel does not tile the toy `_MODEL` shapes. This is the same assertion at
-    module level and at the real Flash shapes, and it is not implied by its neighbour above, which
-    compares two runs of the same path and would pass unchanged if both left a parameter at zero.
-
-    The call count is load-bearing rather than decoration: `dsv4_sparse_attn` raises today
-    instead of falling back, but a regression that reintroduced a fallback would leave this
-    asserting a property of the gather reference.
-    """
-    module = flash_attention(_FLASH_CSA_LAYER, dtype=torch.bfloat16, dsv4_attn="kernel")
-    packed = _packed_context(_KERNEL_DOC_LENS, torch.bfloat16, _flash_config())
-    with torch.device("cuda"):
-        hidden_states = torch.randn(1, sum(_KERNEL_DOC_LENS), _FLASH_MODEL["hidden_size"], dtype=torch.bfloat16)
-    hidden_states.requires_grad_(True)
-
-    calls = []
-    real_kernel = dsv4_attention.dsv4_sparse_attn
-
-    def counting_kernel(q, kv_buf, indices, sinks, scale):
-        calls.append(q.shape[1])
-        return real_kernel(q, kv_buf, indices, sinks, scale)
-
-    monkeypatch.setattr(dsv4_attention, "dsv4_sparse_attn", counting_kernel)
-
-    output, _ = module(hidden_states, packed=packed)
-    assert calls == [sum(_KERNEL_DOC_LENS)], f"the forward never reached the kernel, calls={calls}"
-    with torch.device("cuda"):
-        weight = torch.randn_like(output)
-    (output * weight).sum().backward()
-
-    dead, unexpectedly_alive = [], []
-    for name, param in module.named_parameters():
-        has_grad = param.grad is not None and param.grad.norm().item() > 0
-        # The same expectation `test_deepseek_v4_backward` and `_compare_accumulated_grads` carry:
-        # the indexer reaches the loss only through integer top-k indices, so nothing
-        # differentiates back into it.
-        if ".indexer." in name:
-            if has_grad:
-                unexpectedly_alive.append(name)
-        elif not has_grad:
-            dead.append(name)
-
-    assert not dead, f"Parameters with zero/no gradients: {dead}"
-    assert not unexpectedly_alive, f"Lightning Indexer parameters received a gradient: {unexpectedly_alive}"
-    assert hidden_states.grad is not None and hidden_states.grad.norm() > 0, (
-        "the hidden states received no gradient, so nothing reached the layer's inputs"
-    )
-
-
-# The HCA layer of the Flash config, which nothing else here builds: every other test at these
-# shapes takes `_FLASH_CSA_LAYER`. Documents are exact multiples of the HCA compress rate of 128,
-# so both own whole entries and only the numbering, and with it the RoPE position, moves. Measured
-# over 20 seeds the worst deviation is 3.0e-6 on the output and 4.6e-6 on a gradient, so
-# `_PACKED_RTOL` holds here with room to spare, as it does for the gather test above.
-_FLASH_HCA_LAYER = 1
-_FLASH_HCA_DOCS = (256, 512)
-
-
-def test_flash_hca_attention_packed_matches_unpacked():
-    """An HCA layer at production shapes must answer each document as if it stood alone.
-
-    HCA has no indexer and no sparse path: `DeepseekV4Attention` sends any layer that is not
-    `compressed_sparse_attention` to the dense `_eager_with_entries` regardless of `attn_impl`,
-    so there is no second implementation to compare against and this asserts self-consistency
-    rather than equivalence. What it covers is the part the toy `_MODEL` cannot reach: a compress
-    rate of 128 over 512 channels, where an entry pools 128 tokens and a document boundary that
-    the compressor failed to respect would pull a whole other document into one entry.
-
-    `test_attention_packed_matches_unpacked[hca]` asserts the same invariant at toy shapes and
-    rate 8. This is the only test that instantiates the Flash config's HCA layer at all.
-    """
-    module = flash_attention(_FLASH_HCA_LAYER, dtype=torch.float32, dsv4_attn="eager")
-    assert module.layer_type == "heavily_compressed_attention", (
-        f"expected the Flash config's HCA layer, got {module.layer_type}"
-    )
-    seq_len = sum(_FLASH_HCA_DOCS)
-    packed_input, alone_input = _flash_hidden_states(seq_len)
-    packed = _packed_context(_FLASH_HCA_DOCS, torch.float32, _flash_config())
-
-    q_residual = module.q_a_norm(module.q_a_proj(packed_input.detach()))
-    _, block_bias = module.compressor(packed_input.detach(), q_residual, packed)
-    assert (block_bias[:, :, _doc_slice(_FLASH_HCA_DOCS, 1)] == 0).any(), (
-        "vacuous probe: no query of the second document reads a compressed entry"
-    )
-
-    packed_output, _ = module(packed_input, packed=packed)
-    with torch.device("cuda"):
-        weight = torch.randn_like(packed_output)
-    (packed_output * weight).sum().backward()
-    packed_grads = _take_grads(module)
-
-    for index, length in enumerate(_FLASH_HCA_DOCS):
-        span = _doc_slice(_FLASH_HCA_DOCS, index)
-        alone_output, _ = module(
-            alone_input[:, span], packed=_packed_context((length,), torch.float32, _flash_config())
-        )
-        _assert_relative(packed_output[:, span], alone_output, _PACKED_RTOL, f"document {index}")
-        (alone_output * weight[:, span]).sum().backward()
-
-    _compare_accumulated_grads(module, packed_grads, rtol=_PACKED_GRAD_RTOL)
-    _assert_relative(alone_input.grad, packed_input.grad, _PACKED_GRAD_RTOL, "hidden states gradient")
